test: split classic battle match flow tests
