+++ tests/helpers/classicBattle/selectionPrompt.test.js
@@ -1,0 +1,97 @@
+import { describe, it, expect, vi, beforeEach, afterEach } from "vitest";
+import { setupClassicBattleDom } from "./utils.js";
+
+vi.mock("../../../src/helpers/motionUtils.js", () => ({
+  shouldReduceMotionSync: () => true
+}));
+vi.mock("../../../src/utils/scheduler.js", () => ({
+  onFrame: (cb) => {
+    const id = setTimeout(() => cb(performance.now()), 16);
+    return id;
+  },
+  onSecondTick: (cb) => {
+    const id = setInterval(() => cb(performance.now()), 1000);
+    return id;
+  },
+  cancel: (id) => {
+    clearTimeout(id);
+    clearInterval(id);
+  },
+  start: vi.fn(),
+  stop: vi.fn()
+}));
+
+let generateRandomCardMock;
+vi.mock("../../../src/helpers/randomCard.js", () => ({
+  generateRandomCard: (...args) => generateRandomCardMock(...args)
+}));
+
+let getRandomJudokaMock;
+let renderMock;
+let JudokaCardMock;
+vi.mock("../../../src/helpers/cardUtils.js", () => ({
+  getRandomJudoka: (...args) => getRandomJudokaMock(...args)
+}));
+vi.mock("../../../src/components/JudokaCard.js", () => {
+  renderMock = vi.fn();
+  JudokaCardMock = vi.fn().mockImplementation(() => ({ render: renderMock }));
+  return { JudokaCard: JudokaCardMock };
+});
+
+let fetchJsonMock;
+vi.mock("../../../src/helpers/dataUtils.js", () => ({
+  fetchJson: (...args) => fetchJsonMock(...args),
+  importJsonModule: vi.fn()
+}));
+
+vi.mock("../../../src/helpers/utils.js", () => ({
+  createGokyoLookup: () => ({})
+}));
+
+let currentFlags;
+vi.mock("../../../src/helpers/featureFlags.js", () => ({
+  featureFlagsEmitter: new EventTarget(),
+  initFeatureFlags: vi.fn().mockResolvedValue({ featureFlags: currentFlags }),
+  isEnabled: (flag) => currentFlags[flag]?.enabled ?? false
+}));
+
+let timerSpy;
+
+beforeEach(() => {
+  ({
+    timerSpy,
+    fetchJsonMock,
+    generateRandomCardMock,
+    getRandomJudokaMock,
+    renderMock,
+    currentFlags
+  } = setupClassicBattleDom());
+});
+
+afterEach(() => {
+  timerSpy.clearAllTimers();
+  vi.restoreAllMocks();
+});
+
+describe("classicBattle selection prompt", () => {
+  it("shows selection prompt until a stat is chosen", async () => {
+    const battleMod = await import("../../../src/helpers/classicBattle.js");
+    const store = battleMod.createBattleStore();
+    battleMod._resetForTest(store);
+    await battleMod.startRound(store);
+    expect(document.querySelector(".snackbar").textContent).toBe("Select your move");
+    timerSpy.advanceTimersByTime(5000);
+    expect(document.querySelector(".snackbar")).toBeNull();
+    expect(document.querySelector("header #round-message").textContent).toBe("");
+    document.getElementById("player-card").innerHTML =
+      `<ul><li class="stat"><strong>Power</strong> <span>5</span></li></ul>`;
+    document.getElementById("computer-card").innerHTML =
+      `<ul><li class="stat"><strong>Power</strong> <span>3</span></li></ul>`;
+    {
+      const p = battleMod.handleStatSelection(store, "power");
+      await vi.runAllTimersAsync();
+      await p;
+    }
+    expect(document.querySelector(".snackbar")?.textContent).not.toBe("Select your move");
+  });
+});
